Build script for hopefully soon multi-browser support & Update README [Build]
This initial version of this build script should allow me to work on the same code, just for any differences that have to be made, to just do automatically. While in a perfect world I could wish that Firefox and Chrome-based browsers are united. Alas, this can't really happen. This extension was built for Firefox first, other browsers second. I just recently tested starting this extension on Edge, but to no avail. Hopefully soon I can make them work on multiple browsers. Most of this extension is done by now, so I guess this could become a priority for me.