Optimise observer [YouTube Music]
Used Deepseek R1 AI for this one, the code looks cleaner and honestly way better, I will never be able to code this way... Anyway, the main problem was that the observer was firing way too often (especially when adjusting volume.. yikes). Now it doesn't and it works smoothly now.